fix integer-type and other compiler warnings across core
Minimal-blast-radius fixes: loop variable types int->size_t where bounds
are size_t, explicit static_casts at signedness boundaries, constructor
initializer-list reordering, commented-out unused parameter names, and
range-loop-construct fixes. No member types, signatures, or return
types changed (Degree() keeps its int/-1 sentinel).

+++ core/test/classes/node_serialization_test.cpp
@@ -354,7 +354,7 @@ BOOST_AUTO_TEST_CASE(system_clone)
 	BOOST_CHECK_EQUAL(variables1.size(), variables2.size());
 
 
-	for (int ii=0; ii<variables2.size(); ++ii)
+	for (size_t ii=0; ii<variables2.size(); ++ii)
 	{
 		BOOST_CHECK(variables1[ii].get() != variables2[ii].get());
 	}
